update. add manifest json
