+++ src/utils/storageUtils.js
@@ -1,0 +1,7 @@
+export function getUserName() {
+    return localStorage.getItem('userName');
+}
+
+export function setUserName(name) {
+    localStorage.setItem('userName', name);
+}
